chore: move parseURL to utils
